Future: ROM Interface library
Just a few small changes, including a reordering of modules. The exact
order of files to be included in RomInterface.Lib is not clear because
so few modules made it into the final ROM. Even worse, I cannot find any
release of Interface.o that matches the updates made here. So these
changes are enough to round-trip, but they do not reflect the true Apple
sources very well.

The nasty PBxxx(..., async) calls have been replaced with proper
PBxxxSync/PBxxxAsync calls. This allows trap inlining to work correctly.